Create separate folders for S3 and strategies SN-603
