+++ demo/Saritasa.NetForge.Demo/Infrastructure/Admin/ShopAdminConfiguration.cs
@@ -1,4 +1,5 @@
 ﻿using Saritasa.NetForge.Demo.Infrastructure.UploadFiles;
+using Saritasa.NetForge.Demo.Infrastructure.UploadFiles.Strategies;
 using Saritasa.NetForge.Demo.Models;
 using Saritasa.NetForge.Domain.Enums;
 using Saritasa.NetForge.DomainServices;
