+++ demo/Saritasa.NetForge.Demo/Infrastructure/DependencyInjection/NetForgeModule.cs
@@ -2,7 +2,7 @@
 using Saritasa.NetForge.Demo.Constants;
 using Saritasa.NetForge.Demo.Infrastructure.Admin;
 using Saritasa.NetForge.Demo.Infrastructure.Extensions;
-using Saritasa.NetForge.Demo.Infrastructure.UploadFiles;
+using Saritasa.NetForge.Demo.Infrastructure.UploadFiles.S3Storage;
 using Saritasa.NetForge.Demo.Models;
 using Saritasa.NetForge.Domain.Entities.Options;
 using Saritasa.NetForge.Infrastructure.Abstractions.Interfaces;
