+++ demo/Saritasa.NetForge.Demo/Infrastructure/UploadFiles/Strategies/UploadFileToS3Strategy.cs
@@ -1,8 +1,9 @@
 ﻿using Microsoft.AspNetCore.Components.Forms;
+using Saritasa.NetForge.Demo.Infrastructure.UploadFiles.S3Storage;
 using Saritasa.NetForge.Domain;
 using Saritasa.NetForge.Infrastructure.Abstractions.Interfaces;
 
-namespace Saritasa.NetForge.Demo.Infrastructure.UploadFiles;
+namespace Saritasa.NetForge.Demo.Infrastructure.UploadFiles.Strategies;
 
 /// <summary>
 /// Strategy to upload file to S3.
